Technical documentation Page
Technical documentation refers to any document that explains the use, functionality, creation, or architecture of a product. Think of it as a nuts-and-bolts “how to” guide for a users, new hires, administrators, and anyone else who needs to know how your product works, I built it with HTML and Css.